Abstracted TabBar into own component
