+++ dinnit/components/TabBar.jsx
@@ -1,0 +1,104 @@
+import { View, Text, Image } from "react-native";
+import React from "react";
+import { Tabs } from "expo-router";
+import icons from "../constants/icons"; // Default import here
+
+const TabIcon = ({ icon, color, name, focused }) => {
+  return (
+    <View className="items-center justify-center gap-2">
+      <Image
+        source={icon}
+        resizeMode="contain"
+        style={{ tintColor: color }} // Applied style for tintColor
+        className="w-6 h-6"
+      />
+      <Text
+        className={`${focused ? "font-semibold" : "font-normal"} text-xs`}
+        style={{ color }}
+      >
+        {name}
+      </Text>
+    </View>
+  );
+};
+
+const TabBar = () => {
+  return (
+    <Tabs
+      screenOptions={{
+        tabBarShowLabel: false,
+        tabBarActiveTintColor: "#FF4C4C",
+        tabBarInactiveTintColor: "#CDCDE0",
+        tabBarStyle: {
+          backgroundColor: "#333",
+          borderTopWidth: 1,
+          borderTopColor: "#232533",
+          height: 84,
+        },
+      }}
+    >
+      <Tabs.Screen
+        name="home"
+        options={{
+          title: "Home",
+          headerShown: false,
+          tabBarIcon: ({ color, focused }) => (
+            <TabIcon
+              icon={icons.home} // Accessed icons via default import
+              color={color}
+              name="Home"
+              focused={focused}
+            />
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="discover"
+        options={{
+          title: "Discover",
+          headerShown: false,
+          tabBarIcon: ({ color, focused }) => (
+            <TabIcon
+              icon={icons.eye} // Accessed icons via default import
+              color={color}
+              name="Discover"
+              focused={focused}
+            />
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="search"
+        options={{
+          title: "Search",
+          headerShown: false,
+          tabBarIcon: ({ color, focused }) => (
+            <TabIcon
+              icon={icons.search} // Accessed icons via default import
+              color={color}
+              name="Search"
+              focused={focused}
+            />
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="profile"
+        options={{
+          title: "Profile",
+          headerShown: false,
+          tabBarIcon: ({ color, focused }) => (
+            <TabIcon
+              icon={icons.profile} // Accessed icons via default import
+              color={color}
+              name="Profile"
+              focused={focused}
+            />
+          ),
+        }}
+      />
+    </Tabs>
+  );
+};
+
+export default TabBar;
